правка типов при фильтрации

+++ include/IPV4Pool.h
@@ -11,13 +11,13 @@ private:
 	static const int8_t mEmptyFilter = -1;
 	std::list<IPV4> mStorage;
 private:
-	bool checkFilter(const IPV4& addr, const std::vector<int8_t> & values);
-	bool checkFilterAny(const IPV4& addr, int8_t value);
+	bool checkFilter(const IPV4& addr, const std::vector<int> & values);
+	bool checkFilterAny(const IPV4& addr, uint8_t value);
 public:
 	void readFromFile(const std::string & fileName);
 	IPV4Pool filter(const std::string& mask);
-	IPV4Pool filterAny(int8_t value);
-	IPV4Pool filter(int8_t first, int8_t second = mEmptyFilter, int8_t third = mEmptyFilter, int8_t fourth = mEmptyFilter);
+	IPV4Pool filterAny(uint8_t value);
+	IPV4Pool filter(int first, int second = mEmptyFilter, int third = mEmptyFilter, int fourth = mEmptyFilter);
 	void lexicSort();
 	friend std::istream& operator>> (std::istream& in,  IPV4Pool& point);
 	friend std::ostream& operator<< (std::ostream& out, const IPV4Pool& point);
